审查修复一批正确性与体验问题；补齐 README 与 requirements
- 高 DPI 下 save_size 按 CSS 像素保存导致窗口宽度逐次缩水：改按物理像素
- 单实例互斥改用 use_last_error（直接调 GetLastError 读数不可靠）
- 搜索框内有选中文字时不再被 Ctrl+C 劫持为复制路径
- 托盘启动在页面加载完成前呼出会被 _on_loaded 藏回：加 _summoned_early 保护
- 打开/定位成功后自动隐藏置顶窗口；frecency 改为动作成功后记录并加线程锁
- 热键注册失败时前端横幅提示；hotkey.stop 消息队列竞态修复
- window.hwnd 改用 ToInt64 防句柄溢出；清理 search 内遮蔽的 import time
- install/uninstall.ps1 不再给自动变量 $args 赋值
- 新增 README.md 与 requirements.txt

+++ glance/window.py
@@ -167,7 +167,7 @@ class NativeWindow:
 
     # ---- 基础 ----
     def hwnd(self):
-        return int(self._w.native.Handle.ToInt32())
+        return int(self._w.native.Handle.ToInt64())
 
     def ui_invoke(self, fn):
         try:
